requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
blessed==1.14.2
botocore==1.10.38
cached-property==1.4.2
cement==2.8.2
certifi==2018.4.16
cffi==1.11.5
cftime==1.0.0
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
colorama==0.3.9
cryptography==2.2.1
cycler==0.10.0
Cython==0.28.3
descartes==1.1.0
docker==3.3.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Fiona==1.7.12
Flask==1.0.2
Flask-RESTful==0.3.6
GDAL==2.2.4
geopandas==0.3.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
munch==2.3.2
netCDF4==1.4.0
numpy==1.14.5
pandas==0.23.1
pathspec==0.5.5
polyline==1.3.2
psycopg2==2.7.5
pycparser==2.18
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyproj==1.9.5.1
PySAL==1.14.3
PySocks==1.6.8
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.4
requests==2.18.4
Rtree==0.8.3
schedule==0.5.0
scipy==0.19.1
semantic-version==2.5.0
Shapely==1.6.4.post1
six==1.11.0
SQLAlchemy==1.2.8
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
tornado==5.0.2
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.48.0
Werkzeug==0.14.1
xarray==0.10.7
